Add song data fetching from JSON file and include album cover
In this commit, we have updated the way we fetch and display the song data in our Music Player application. Instead of using hardcoded data in the HTML file, we now fetch the song data from a JSON file (song.json) which includes the song title, artist, album, file name, and album cover image. This allows for a more dynamic and maintainable approach to handling song data.

We have also updated the JavaScript code to fetch the data from the JSON file and display the song list accordingly. The album cover images are now part of the song data, making it easier to associate the cover image with the corresponding song.

Please make sure to include the appropriate album cover image files in the music-player-custom/assets/ folder.