Add shuffle feature and fix previous button behavior
This commit introduces a new shuffle feature to the music player. A shuffle button has been added next to the previous button, allowing the user to enable or disable the shuffle mode. When shuffle mode is active, the next song will be selected randomly from the playlist.

Additionally, this commit fixes a bug related to the previous button behavior. Previously, when the shuffle mode was active and the user pressed the previous button, the application would play the song with the previous index. Now, with the shuffle mode active, pressing the previous button will return to the song that was playing before the shuffle mode was enabled. This behavior is more in line with user expectations.